Binding Service across spaces is not permitted
When deploying the monitoring in a different space to the application,
the redis exporter would not bind across CF spaces.
Since it has no need to be bound, the URL and credentials are passed
in as Environment variables, this can be removed.
And it neatly keeps the exporter in a different space to the application.